fix(dynamic-agents): rename logging.py to avoid stdlib collision
- Rename logging.py to log_config.py to prevent shadowing Python's
  built-in logging module which caused AttributeError on startup
- Remove CORS_ORIGINS env var from docker-compose to use Pydantic default
